cleanup: isolate node database rebuild slot
